meme-feed/requirements.txt
asgiref==3.7.2
blinker==1.4
certifi==2023.7.22
charset-normalizer==3.3.1
command-not-found==0.3
crispy-bootstrap4==2023.1
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.2.6
django-crispy-forms==2.1
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
Pillow==10.1.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.8.0
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.0.7
wadllib==1.3.6
zipp==1.0.0
